Fixed subtile bug in the cleanup routine which deleted all the incremental backups till there was a daily backup ouside the daily retention
There the 'newest daily outside retention' was still undef, an undef was put in
the keep_dailys array. is_in_array fails in finding an undef in the array.
This 'Fix' only removes the trigger which inserted an undef into the array.
The function is_in_array still fails on an undef in the array



git-svn-id: https://dev.snt.utwente.nl/svn/snt-syscom/snt-backup/trunk@469 53864893-6df6-0310-a35a-92b7ff90ca16